Move positioner to its own file
