+++ src/dodal/devices/i10/diagnostics.py
@@ -1,6 +1,4 @@
-from bluesky.protocols import Movable
 from ophyd_async.core import (
-    AsyncStatus,
     Device,
     StandardReadable,
     StrictEnum,
@@ -13,7 +11,6 @@ from ophyd_async.epics.core import (
     epics_signal_r,
     epics_signal_rw,
 )
-from ophyd_async.epics.motor import Motor
 
 from dodal.devices.current_amplifiers import (
     CurrentAmpDet,
@@ -23,6 +20,7 @@ from dodal.devices.current_amplifiers import (
     FemtoDDPCA,
     StruckScaler,
 )
+from dodal.devices.positioner import Positioner
 
 
 class D3Position(StrictEnum):
@@ -68,36 +66,6 @@ class InOutReadBackTable(StrictEnum):
     IN_BEAM = "In Beam"
     FAULT = "Fault"
     OUT_OF_BEAM = "Out of Beam"
-
-
-class Positioner(StandardReadable, Movable):
-    """1D stage with a enum table to select positions."""
-
-    def __init__(
-        self,
-        prefix: str,
-        positioner_enum: type[StrictEnum],
-        positioner_suffix: str = "",
-        Positioner_pv_suffix: str = ":MP:SELECT",
-        name: str = "",
-    ) -> None:
-        self._stage_motion = Motor(prefix=prefix + positioner_suffix)
-        with self.add_children_as_readables(Format.CONFIG_SIGNAL):
-            self.stage_position = epics_signal_rw(
-                positioner_enum,
-                read_pv=prefix + positioner_suffix + Positioner_pv_suffix,
-            )
-        super().__init__(name=name)
-        self.positioner_enum = positioner_enum
-
-    @AsyncStatus.wrap
-    async def set(self, value: StrictEnum) -> None:
-        if value in self.positioner_enum:
-            await self.stage_position.set(value=value)
-        else:
-            raise ValueError(
-                f"{value} is not an allow position. Position must be: {self.positioner_enum}"
-            )
 
 
 class I10PneumaticStage(StandardReadable):
